# speedup arm int functions?