I was looking at this curious little guy which seems close to what we want? https://github.com/openjdk/jdk/blob/b3fec6b5f32c338ae1a84dd20bdcbd3d9b7186f3/src/hotspot/cpu/aarch64/aarch64_vector_ad.m4#L1375-L1390

Not sure what it is there for, maybe to speedup some hashing function or crypto or something.

it has a test: https://github.com/openjdk/jdk/blob/b3fec6b5f32c338ae1a84dd20bdcbd3d9b7186f3/test/hotspot/jtreg/compiler/loopopts/superword/Vec_MulAddS2I.java#L81

So i changed an integer function to look more like that and it shows speedups on arm without slowdown on x86. needs more investigation. maybe can be applied to some of the other functions, not sure, i am sleepy.

```
master:
Benchmark                                   (size)   Mode  Cnt  Score   Error   Units
VectorUtilBenchmark.binaryDotProductScalar    1024  thrpt    5  3.121 ± 0.001  ops/us
VectorUtilBenchmark.binaryDotProductVector    1024  thrpt    5  6.742 ± 0.037  ops/us

patch:
Benchmark                                   (size)   Mode  Cnt  Score   Error   Units
VectorUtilBenchmark.binaryDotProductScalar    1024  thrpt    5  4.523 ± 0.077  ops/us
VectorUtilBenchmark.binaryDotProductVector    1024  thrpt    5  6.682 ± 0.066  ops/us
```