@rmuir 2023-11-01
thank you for the explanation, so now it just adds another mystery for me.  Your x64 is still i5-11400 ? I will look into this more.

Yes on the ARM, the assembly didn't look like what i wanted but inputs are `byte` and not `short` which might be relevant. But performance improved, so that means we should investigate wtf is happening and try to take it to the limit :)