@rmuir 2023-11-01
and what is the compiler's excuse this time? :) no floating point here. I should just be able to write a simple loop and get good performance!