@ChrisHegarty 2023-11-01
> Your x64 is still i5-11400 ? I will look into this more.

Yes.