@rmuir 2023-11-01
thanks @ChrisHegarty ! I will try to dig in more to this. Again i don't see speedups on my x86 so something has changed (maybe architecture) thats allowing more to happen in parallel on your CPU or something. wikichip has been down so I don't know any answers yet. nor if it can apply to the vector case.