@rmuir 2023-11-03
equivalent on intel ice lake: https://www.felixcloutier.com/x86/vpdpbusd
IMO, we should figure out a path to using these, to get the best performance from the binary vectors. it isn't useful to spend time trying to do many other hacks around the situation when the "problem has been solved", we just have to make use of the hardware.

but i think this little patch won't hurt anyone and improves the scalar baseline.