@rmuir 2023-11-06
When looking at `SDOT` to do this, i was able to accomplish it with another vector API, just as basis for comparison: 

https://godbolt.org/z/9cv5WaGaT

Even if you lower `-march=armv8-a` to exclude `SDOT`, it just uses `smull`/`smull2` and `saddw`/`saddw2` to work on lower/upper halves of vector.

On the other hand here I am struggling with crazy conversions and generics part numbers and slicing abstractions :( I don't actually want to convert anything!

The only lone use of long multiplication i find in openjdk arm support is for the "curious little guy".