@rmuir 2023-11-06
we should test more. I can test on more machines later too. Maybe good motivation to get full automation going for doing that.